Worked on review comments
