+++ drivers/unifiedPlatform/pdsLibs/induceErrorsLib.go
@@ -1,0 +1,18 @@
+package pdslibs
+
+import (
+	"fmt"
+	"github.com/portworx/torpedo/pkg/log"
+)
+
+func PerformMultiplication() error {
+	log.InfoD("Error Injection begins here .. ")
+	mul := 1
+	for i := 1; i <= 10; i++ {
+		mul *= i
+		log.InfoD("Multiplication is [%v]", mul)
+	}
+	fmt.Println("Multiplication of the first 10 integers:", mul)
+	log.InfoD("Error Injection is completed")
+	return nil
+}
