+++ drivers/unifiedPlatform/stworkflows/workflowResiliency.go
@@ -16,8 +16,8 @@ func (wkflwResi *WorkflowResiliency) MarkResiliencyTC(resiliency bool) {
 	wkflwResi.ResiliencyFlag = resiliency
 	log.InfoD("Execution of a Resiliency TestCase Begins ...")
 }
-func (wkflwResi *WorkflowResiliency) InduceFailureAndExecuteResiliencyScenario(namespace string, failureType string, method error) error {
-	err := resiLibs.InduceFailureAfterWaitingForCondition(namespace, failureType, method, wkflwResi.ResiliencyFlag)
+func (wkflwResi *WorkflowResiliency) InduceFailureAndExecuteResiliencyScenario(namespace string, failureType string) error {
+	err := resiLibs.InduceFailureAfterWaitingForCondition(namespace, failureType, wkflwResi.ResiliencyFlag)
 	if err != nil {
 		return err
 	}
